Move Encryption implementation > UserSecretStorage
They are only used by the UserSecretStorage solution, so moving them to
that 'namespace' makes perfect sense
