+++ library/tiqr/Tiqr/UserSecretStorage/UserSecretStorageTrait.php
@@ -1,0 +1,1 @@
+<?php
