+++ library/tiqr/Tiqr/UserStorage/FileTrait.php
@@ -1,0 +1,1 @@
+<?php
